requirements.txt
django
djangorestframework
debugpy
psycopg2-binary
celery
redis
urllib3
aiohttp-retry
requests
selenium
django-celery-beat